requirements.txt
absl-py==2.1.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
black==24.8.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
contourpy==1.1.1
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9604f5995cc628619f0e4fd913453b4d7d61db3f
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.55.8
fsspec==2025.2.0
fvcore==0.1.5.post20221221
google-auth==2.38.0
google-auth-oauthlib==1.0.0
grpcio==1.70.0
hydra-core==1.3.2
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.9
Jinja2==3.1.5
kiwisolver==1.4.7
lazy_loader==0.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
packaging==24.2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
portalocker==3.0.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.7
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.21.0
scipy==1.10.1
six==1.17.0
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tifffile==2023.7.10
tomli==2.2.1
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.6
yacs==0.1.8
zipp==3.20.2
